feat: make answering machine detection by twilio asynchronous
A dictionary of active `CallManager` instances with call SID keys is stored as a class attribute. That way a separate `/amdstatus` route handler can get the call manager for a given call SID and abort the call if the Twilio AMD determined an answering machine or fax to be on the other end.
A separate `AnsweringMachineDetected` exception now exists for this purpose.